refactor: Use more Go types in IL2P/FX.25/HDLC code

+++ src/hdlc_rec.go
@@ -448,8 +448,8 @@ func hdlc_rec_bit_new(channel int, subchannel int, slice int, _raw int, is_scram
 	// Don't waste time on this if AIS.  EAS does not get this far.
 
 	if g_audio_p.achan[channel].modem_type != MODEM_AIS {
-		fx25_rec_bit(C.int(channel), C.int(subchannel), C.int(slice), C.int(IfThenElse(dbit, 1, 0)))
-		il2p_rec_bit(C.int(channel), C.int(subchannel), C.int(slice), C.int(IfThenElse(raw, 1, 0))) // Note: skip NRZI.
+		fx25_rec_bit(channel, subchannel, slice, IfThenElse(dbit, 1, 0))
+		il2p_rec_bit(channel, subchannel, slice, IfThenElse(raw, 1, 0)) // Note: skip NRZI.
 	}
 
 	/*
